app/requirements.txt
Flask==2.3.2
Flask-Migrate>=3.1.0
Flask-SQLAlchemy==3.0.3
python-dateutil>=2.8.2
psycopg2-binary==2.9.6

